+++ ts-sdk/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @crypticdot/fusionamm-core
 
+## 1.0.56
+
 ## 1.0.55
 
 ## 1.0.54